Major rework to have the GUI and model handle all the various actions appropriately for major, minor and patch level libraries.
Known problems are capturing documentation changes in minor and patches.
The GUI controls are correct but the model does not save changes. The
other problem is that enumeration literals from earlier versions are not
shown as inherited properties.

There are some junit failures, but these appear to be due to the test
and not the code. More testing and fixing the tests to come.

With the amount of refactoring done, there should be some more detailed
testing before a major release is done.